requirements.txt
dash==2.18.2
dash-bootstrap-components==1.6.0
plotly==5.24.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.3
python-dateutil==2.9.0
anthropic==0.40.0
sqlalchemy
